✨ Basic gradient support
